+++ moVirt/src/main/java/org/ovirt/mobile/movirt/rest/client/requestfactory/verifier/ListHostnameVerifier.java
@@ -1,4 +1,4 @@
-package org.ovirt.mobile.movirt.rest;
+package org.ovirt.mobile.movirt.rest.client.requestfactory.verifier;
 
 import org.androidannotations.annotations.EBean;
